fix bug and add corresponding tests - add dtype to return value

+++ udao/udao/data/tests/iterators/test_query_graph_iterator.py
@@ -83,7 +83,7 @@ class TestQueryGraphIterator:
 
     def test_get_graph(self, sample_iterator: QueryPlanIterator) -> None:
         graph, meta = sample_iterator._get_graph_and_meta("a")
-        assert th.equal(meta, th.tensor([0, 10]))
+        assert th.equal(meta, th.tensor([0, 10], dtype=th.float32))
         assert isinstance(graph, dgl.DGLGraph)
         assert graph.number_of_nodes() == 2
         assert th.equal(
